+++ src/screens/AnalysisGroups/scenes/GroupDetail/index.tsx
@@ -1,7 +1,10 @@
 import * as React from 'react';
+import { Row, Col } from 'react-bootstrap';
 
 import { AnalysisGroupType } from '../../../../services/api/models/analysisGroup';
 import { analysisGroups } from '../../../../services/api/mocks/analysisGroups';
+
+import SampleSimilarity from '../../../../components/SampleSimilarity';
 
 interface Props {
     groupName: string;
@@ -19,10 +22,15 @@ class AnalysisGroupDetail extends React.Component<Props, {}> {
 
   render() {
     return (
-      <div>
+      <Row>
         <h1>{this.group.name}</h1>
         <p>{this.group.description}</p>
-      </div>
+        <Row>
+          <Col lg={12}>
+            <SampleSimilarity />
+          </Col>
+        </Row>
+      </Row>
     );
   }
 }